app/src/requirements.txt
APScheduler==3.6.3
ares==0.7.1
astroid==2.4.2
autopep8==1.5.4
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
click==7.1.2
cryptography==3.3.1
decorator==4.4.2
Flask==1.1.2
idna==2.10
iocextract==1.13.1
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
OpenVulnQuery==1.30
pycodestyle==2.6.0
pycparser==2.20
pycvesearch==1.0.1
pylint==2.6.0
python-telegram-bot==13.1
pytz==2020.5
PyYAML==5.3.1
regex==2020.11.13
requests==2.25.0
schedule==0.6.0
six==1.15.0
toml==0.10.2
tornado==6.1
tzlocal==2.1
urllib3==1.26.2
vulners==1.5.9
Werkzeug==1.0.1
wrapt==1.12.1